+++ src/mba.modulo1.blog.api/Controllers/PostController.cs
@@ -2,31 +2,42 @@
 using mba.modulo1.blog.domain.Entities;
 using MBA.Modulo1.Blog.API.DTO;
 using MBA.Modulo1.Blog.Domain.Interfaces;
+using MBA.Modulo1.Blog.Domain.Services;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Net;
 
 namespace MBA.Modulo1.Blog.API.Controllers;
 
+[Authorize]
 [Route("api/posts")]
 public class PostController : MainController
 {
     private readonly IPostService _postService;
+    private readonly ICommentService _commentService;
     private readonly IPostRepository _postRepository;
     private readonly IMapper _mapper;
 
-    public PostController(IPostService postService, IPostRepository postRepository, IMapper mapper, INotifier notifier) : base(notifier)
+    public PostController(IPostService postService, 
+        ICommentService commentService, 
+        IPostRepository postRepository, 
+        IMapper mapper, 
+        INotifier notifier) : base(notifier)
     {
         _postService = postService;
+        _commentService = commentService;
         _postRepository = postRepository;
         _mapper = mapper;
     }
 
+    [AllowAnonymous]
     [HttpGet]
     public async Task<IEnumerable<PostDTO>> GetAll()
     {
         return _mapper.Map<IEnumerable<PostDTO>>(await _postRepository.GetAllAsync());
     }
 
+    [AllowAnonymous]
     [HttpGet("GetById/{id:guid}")]
     public async Task<ActionResult<PostDTO>> GetByIdAsync(Guid id)
     {
@@ -37,6 +48,7 @@ public class PostController : MainController
         return Ok(post);
     }
 
+    [AllowAnonymous]
     [HttpGet("GetByAuthorId/{id:guid}")]
     public async Task<IEnumerable<PostDTO>> GetAll(string id)
     {
@@ -47,19 +59,20 @@ public class PostController : MainController
     public async Task<ActionResult<PostSaveDTO>> AddAsync(PostSaveDTO postDTO)
     {
         if (!ModelState.IsValid) return CustomResponse(ModelState);
-
-        await _postRepository.AddAsync(_mapper.Map<Post>(postDTO));
+        var post = _mapper.Map<Post>(postDTO);
+        post.AuthorId = GetLoggedUser();
+        await _postRepository.AddAsync(post);
         return CustomResponse(HttpStatusCode.Created, postDTO);
     }
 
     [HttpPut("Update/{id:guid}")]
     public async Task<IActionResult> UpdateAsync(Guid id, PostSaveDTO postDTO)
     {
-
         if (!ModelState.IsValid) return CustomResponse(ModelState);
 
         var post = await GetPostByIdAsync(id);
         if (post == null) return NotFound();
+        if (!UserHasPermition(post.AuthorId)) return Forbid();
 
         //post.UpdatedAt = DateTime.Now;
         post.Title = postDTO.Title;
@@ -74,7 +87,9 @@ public class PostController : MainController
     {
         var post = await GetPostByIdAsync(id);
         if (post == null) return NotFound();
+        if (!UserHasPermition(post.AuthorId)) return Forbid();
 
+        await _commentService.DeleteCommentsbyPostIdAsync(id);
         await _postService.DeleteAsync(id);
         return CustomResponse(HttpStatusCode.NoContent);
     }
